Split code to simplify dep-check rules
